Remove start up syncer
