+++ stdlib/ballerina-http/src/main/java/org/ballerinalang/net/http/HttpConnectorPortBindingListener.java
@@ -18,13 +18,11 @@
 
 package org.ballerinalang.net.http;
 
-import org.ballerinalang.net.http.util.ConnectorStartupSynchronizer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.wso2.transport.http.netty.contract.PortBindingEventListener;
 
 import java.io.PrintStream;
-import java.net.BindException;
 
 /**
  * An implementation of the LifeCycleEventListener. This can be used to listen to the HTTP connector life cycle events.
@@ -36,15 +34,6 @@ public class HttpConnectorPortBindingListener implements PortBindingEventListene
     private static final Logger log = LoggerFactory.getLogger(HttpConnectorPortBindingListener.class);
     private static final PrintStream console = System.out;
 
-    private ConnectorStartupSynchronizer connectorStartupSynchronizer;
-    private String serverConnectorId;
-
-    public HttpConnectorPortBindingListener(ConnectorStartupSynchronizer connectorStartupSynchronizer,
-                                            String serverConnectorId) {
-        this.connectorStartupSynchronizer = connectorStartupSynchronizer;
-        this.serverConnectorId = serverConnectorId;
-    }
-
     @Override
     public void onOpen(String serverConnectorId, boolean isHttps) {
         if (isHttps) {
@@ -52,7 +41,6 @@ public class HttpConnectorPortBindingListener implements PortBindingEventListene
         } else {
             console.println("ballerina: started HTTP/WS server connector " + serverConnectorId);
         }
-        connectorStartupSynchronizer.addServerConnector(serverConnectorId);
     }
 
     @Override
@@ -67,9 +55,5 @@ public class HttpConnectorPortBindingListener implements PortBindingEventListene
     @Override
     public void onError(Throwable throwable) {
         log.error("Error in http server connector", throwable);
-
-        if (throwable instanceof BindException) {
-            connectorStartupSynchronizer.addException(serverConnectorId, (BindException) throwable);
-        }
     }
 }
